prepare .net 5 upgrae

+++ Core/Entities/Video.cs
@@ -4,12 +4,12 @@ namespace Core.Entities
 {
     public class Video : DisplayItem
     {
-        public int Views { get; set; }
-        public int Dislikes { get; set; }
-        public int PublishedAt { get; set; }
-        public int Likes { get; set; }
+        public ulong Views { get; set; }
+        public ulong Dislikes { get; set; }
+        public uint PublishedAt { get; set; }
+        public ulong Likes { get; set; }
         public string ChannelID { get; set; }
-        public int Comments { get; set; }
+        public ulong Comments { get; set; }
         public bool GeoChecked { get; set; }
         public List<string> Tags { get; set; }
         public Channel Channel { get; set; }
